refactor: move view modules into role packages
Relocate the view bodies from oauth2_provider/views/ into the role packages so
the source tree tells a contributor which side they are on:

- authorization_server/views/: base, introspect, device,
  dynamic_client_registration, application, token
- authorization_server/oidc/views.py: discovery, JWKS, userinfo, RP-Initiated Logout
- resource_server/views/: generic (protected-resource views)

Each old oauth2_provider/views/<name>.py becomes a sys.modules-alias shim
(DeprecationWarning). oauth2_provider/views/__init__.py stays as the
backward-compatible flat aggregator but now imports from the canonical homes, and
the role facades are repointed there too, so importing the library or its URLconf
emits no deprecation warnings. IntrospectTokenView keeps its (now explicit)
cross-role dependency on the resource-server ClientProtectedScopedResourceView.

Full non-e2e suite passes (1033); no first-party warnings on URLconf/view import;
view submodule shims alias with identical object identity; makemigrations --check
reports no changes; ruff clean.
