+++ oauth2_provider/authorization_server/oidc/__init__.py
@@ -5,7 +5,7 @@ by role, e.g.::
 
     from oauth2_provider.authorization_server.oidc import UserInfoView
 
-The view bodies remain in :mod:`oauth2_provider.views.oidc`; names are loaded
+The view bodies live in :mod:`oauth2_provider.authorization_server.oidc.views`; names are loaded
 lazily (PEP 562) to avoid touching the app registry at import time.
 """
 
@@ -13,10 +13,10 @@ import importlib
 
 
 _LAZY = {
-    "ConnectDiscoveryInfoView": "oauth2_provider.views.oidc",
-    "JwksInfoView": "oauth2_provider.views.oidc",
-    "UserInfoView": "oauth2_provider.views.oidc",
-    "RPInitiatedLogoutView": "oauth2_provider.views.oidc",
+    "ConnectDiscoveryInfoView": "oauth2_provider.authorization_server.oidc.views",
+    "JwksInfoView": "oauth2_provider.authorization_server.oidc.views",
+    "UserInfoView": "oauth2_provider.authorization_server.oidc.views",
+    "RPInitiatedLogoutView": "oauth2_provider.authorization_server.oidc.views",
 }
 
 __all__ = sorted(_LAZY)
