+++ oauth2_provider/authorization_server/views/application.py
@@ -1,0 +1,87 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.forms.models import modelform_factory
+from django.urls import reverse_lazy
+from django.views.generic import CreateView, DeleteView, DetailView, ListView, UpdateView
+
+from oauth2_provider.authorization_server.forms import ApplicationForm
+from oauth2_provider.models import get_application_model
+
+
+APPLICATION_FIELDS = (
+    "name",
+    "client_id",
+    "client_secret",
+    "hash_client_secret",
+    "client_type",
+    "authorization_grant_type",
+    "redirect_uris",
+    "post_logout_redirect_uris",
+    "allowed_origins",
+    "algorithm",
+)
+
+
+class ApplicationOwnerIsUserMixin(LoginRequiredMixin):
+    """
+    This mixin is used to provide an Application queryset filtered by the current request.user.
+    """
+
+    fields = "__all__"
+
+    def get_queryset(self):
+        return get_application_model().objects.filter(user=self.request.user)
+
+
+class ApplicationRegistration(LoginRequiredMixin, CreateView):
+    """
+    View used to register a new Application for the request.user
+    """
+
+    template_name = "oauth2_provider/application_registration_form.html"
+
+    def get_form_class(self):
+        return modelform_factory(get_application_model(), form=ApplicationForm, fields=APPLICATION_FIELDS)
+
+    def form_valid(self, form):
+        form.instance.user = self.request.user
+        return super().form_valid(form)
+
+
+class ApplicationDetail(ApplicationOwnerIsUserMixin, DetailView):
+    """
+    Detail view for an application instance owned by the request.user
+    """
+
+    context_object_name = "application"
+    template_name = "oauth2_provider/application_detail.html"
+
+
+class ApplicationList(ApplicationOwnerIsUserMixin, ListView):
+    """
+    List view for all the applications owned by the request.user
+    """
+
+    context_object_name = "applications"
+    template_name = "oauth2_provider/application_list.html"
+
+
+class ApplicationDelete(ApplicationOwnerIsUserMixin, DeleteView):
+    """
+    View used to delete an application owned by the request.user
+    """
+
+    context_object_name = "application"
+    success_url = reverse_lazy("oauth2_provider:list")
+    template_name = "oauth2_provider/application_confirm_delete.html"
+
+
+class ApplicationUpdate(ApplicationOwnerIsUserMixin, UpdateView):
+    """
+    View used to update an application owned by the request.user
+    """
+
+    context_object_name = "application"
+    template_name = "oauth2_provider/application_form.html"
+
+    def get_form_class(self):
+        return modelform_factory(get_application_model(), form=ApplicationForm, fields=APPLICATION_FIELDS)
