+++ oauth2_provider/resource_server/__init__.py
@@ -27,16 +27,16 @@ _LAZY = {
     "ResourceServerValidatorMixin": "oauth2_provider.resource_server.validators",
     "validate_resource_as_url_prefix": "oauth2_provider.resource_server.validators",
     "is_valid_resource_uri": "oauth2_provider.resource_server.validators",
-    # Concrete protected-resource views (bodies remain in oauth2_provider.views).
-    "ProtectedResourceView": "oauth2_provider.views.generic",
-    "ScopedProtectedResourceView": "oauth2_provider.views.generic",
-    "ReadWriteScopedResourceView": "oauth2_provider.views.generic",
-    "ClientProtectedResourceView": "oauth2_provider.views.generic",
-    "ClientProtectedScopedResourceView": "oauth2_provider.views.generic",
-    "ProtectedResourceMetadataView": "oauth2_provider.views.generic",
-    "ScopedProtectedResourceMetadataView": "oauth2_provider.views.generic",
-    "ReadWriteScopedProtectedResourceMetadataView": "oauth2_provider.views.generic",
-    "ClientProtectedResourceMetadataView": "oauth2_provider.views.generic",
+    # Concrete protected-resource views.
+    "ProtectedResourceView": "oauth2_provider.resource_server.views.generic",
+    "ScopedProtectedResourceView": "oauth2_provider.resource_server.views.generic",
+    "ReadWriteScopedResourceView": "oauth2_provider.resource_server.views.generic",
+    "ClientProtectedResourceView": "oauth2_provider.resource_server.views.generic",
+    "ClientProtectedScopedResourceView": "oauth2_provider.resource_server.views.generic",
+    "ProtectedResourceMetadataView": "oauth2_provider.resource_server.views.generic",
+    "ScopedProtectedResourceMetadataView": "oauth2_provider.resource_server.views.generic",
+    "ReadWriteScopedProtectedResourceMetadataView": "oauth2_provider.resource_server.views.generic",
+    "ClientProtectedResourceMetadataView": "oauth2_provider.resource_server.views.generic",
     "OAuthProtectedResourceMetadataView": "oauth2_provider.resource_server.views.metadata",
     # Protected-resource mixins.
     "ResourceServerViewMixin": "oauth2_provider.resource_server.mixins",
